+++ src/features/auth/api/index.ts
@@ -1,2 +1,5 @@
 export * from './types';
 export * from './authApi';
+export * from './utils';
+export * from './api.types';
+export * from './auth.types';
