+++ src/features/auth/lib/index.ts
@@ -1,0 +1,2 @@
+export { useAuth } from './useAuth';
+export { useToken } from './useToken';
