+++ src/features/auth/modal/index.ts
@@ -1,1 +1,3 @@
 export { emailVerificationSchema, type FormValuesEmailVerification } from './emailVerificationSchema';
+export * from './signInSchema';
+export { signUpSchema, type FormValues } from './signUpSchema';
