+++ src/features/main-page/ui/HomePostCard/index.ts
@@ -1,1 +1,1 @@
-export { HomePostCard } from './ui/HomePostCard';
+export { HomePostCard } from './ui';
