+++ src/features/main-page/ui/HomePostCard/ui/index.ts
@@ -1,0 +1,1 @@
+export { HomePostCard } from './HomePostCard';
