+++ src/features/main-page/ui/HomePublicPosts/index.ts
@@ -1,1 +1,1 @@
-export { HomePublicPosts } from './ui/HomePublicPosts';
+export { HomePublicPosts } from './ui';
