+++ src/features/main-page/ui/HomePublicPosts/ui/HomePublicPosts.tsx
@@ -3,7 +3,7 @@ import { useAppDispatch, useAppSelector } from '@/app/lib/hooks';
 import { postsApi, useGetPublicPostsQuery } from '@/src/entities/post/api/postsApi';
 import { Post, PostsResponse } from '@/src/entities/post/model/posts.schemas';
 import { HomePostCard } from '@/src/features/main-page/ui/HomePostCard';
-import { useEffect, useRef } from 'react';
+import { useEffect, useMemo, useRef } from 'react';
 import styles from './HomePublicPosts.module.scss';
 
 type Props = {
@@ -12,17 +12,20 @@ type Props = {
 
 export const HomePublicPosts = ({ initialPosts }: Props) => {
   const dispatch = useAppDispatch();
-  const defaultArgs = {
-    endCursorPostId: 0,
-    pageSize: 4,
-    sortBy: 'createdAt',
-    sortDirection: 'desc' as const
-  };
+  const defaultArgs = useMemo(
+    () => ({
+      endCursorPostId: 0,
+      pageSize: 4,
+      sortBy: 'createdAt',
+      sortDirection: 'desc' as const
+    }),
+    []
+  );
   // проверяем кэш, достаем его из стора
   const dataFromCache = useAppSelector((state) => postsApi.endpoints.getPublicPosts.select(defaultArgs)(state).data);
 
   // ref нужен, чтобы гидрировать только один раз, дальше будем скипать
-  const needHydrateStateRef = useRef(!!initialPosts.items && !dataFromCache?.items);
+  const needHydrateStateRef = useRef(Boolean(initialPosts.items?.length) && !dataFromCache?.items.length);
 
   const {
     data: postsData,
@@ -36,14 +39,13 @@ export const HomePublicPosts = ({ initialPosts }: Props) => {
   });
 
   useEffect(() => {
-    if (needHydrateStateRef.current) {
-      needHydrateStateRef.current = false;
-      const thunk = postsApi.util.upsertQueryData('getPublicPosts', {}, initialPosts);
-      dispatch(thunk);
-    }
-  }, [dispatch, initialPosts]);
+    if (!needHydrateStateRef.current) return;
+    needHydrateStateRef.current = false;
+    const thunk = postsApi.util.upsertQueryData('getPublicPosts', defaultArgs, initialPosts);
+    dispatch(thunk);
+  }, [dispatch, initialPosts, defaultArgs]);
 
-  const dataForRender = postsData?.items || initialPosts.items;
+  const dataForRender = postsData?.items ?? dataFromCache?.items ?? initialPosts.items;
 
   if (isFetching && !dataForRender) return <div>Загрузка...</div>;
   if (error) return <div>Ошибка загрузки</div>;
