+++ src/features/main-page/ui/HomePublicPosts/ui/index.ts
@@ -1,0 +1,1 @@
+export { HomePublicPosts } from './HomePublicPosts';
