+++ src/features/main-page/ui/HomeRegisteredUsers/ui/index.ts
@@ -1,0 +1,1 @@
+export { HomeRegisteredUsers } from './HomeRegisteredUsers';
