+++ src/features/main-page/ui/index.ts
@@ -1,0 +1,4 @@
+export { default as MainPage } from './MainPage';
+export * from './HomePublicPosts';
+export * from './HomeRegisteredUsers';
+export * from './HomePostCard';
