+++ src/features/post/delete-post/lib/index.ts
@@ -1,0 +1,1 @@
+export { useDeletePost } from './useDeletePost';
