+++ src/features/posts/lib/index.ts
@@ -1,0 +1,1 @@
+export { useInfinitePosts } from './useInfinitePosts';
