+++ src/widgets/post-modal/index.ts
@@ -1,0 +1,2 @@
+export * from './lib';
+export * from './ui';
