+++ src/widgets/post-modal/lib/index.ts
@@ -1,0 +1,1 @@
+export { getOwnPostMenuItems, getFollowedUserPostMenuItems } from './postMenuItems';
