+++ src/widgets/post-modal/ui/PostModal/PostModal.module.scss
