+++ src/widgets/post-modal/ui/PostModal/index.ts
@@ -1,0 +1,1 @@
+export { PostModal } from './PostModal';
