+++ src/widgets/post-modal/ui/PostModalController/index.ts
@@ -1,0 +1,1 @@
+export { PostModalController } from './PostModalController';
