+++ src/widgets/post-modal/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './PostModal';
+export * from './PostModalController';
