--- src/widgets/post/index.ts
@@ -1,1 +1,0 @@
-export { PostModal } from './ui/PostModal/PostModal';
